Fix assert when viewing legacy user registration lists
This bug doesn't show up on release builds with QT4, I am not sure on
the effect on QT5 builds. When you connect a recent debug build to a
"legacy" server and view the user list, it will crash with the message:

ASSERT: "visual != -1" in file itemviews\qheaderview.cpp, line 1184

This is related to setting resize modes on columns that don't exist.